Rename idp test to federation settings
